+++ src/core/global.d.ts
@@ -12,7 +12,7 @@ declare global {
   interface ErrorConstructor {
     captureStackTrace?(
       targetObject: object,
-      constructorOpt?: (...args: any[]) => any
+      constructorOpt?: (...args: unknown[]) => unknown
     ): void;
   }
 }